futures: use modules to simplify feature-flagging
